wallet: extract GetAccount shared workflow
